nova 13.0.0.0b3 development milestone

meta:version: 13.0.0.0b3
meta:series: mitaka
meta:release-type: development milestone
meta:announce: openstack-announce@lists.openstack.org
meta:pypi: no
meta:first: no

+++ nova/network/linux_net.py
@@ -1339,7 +1339,7 @@ def _set_device_mtu(dev, mtu=None):
                       check_exit_code=[0, 2, 254])
 
 
-def _create_veth_pair(dev1_name, dev2_name):
+def _create_veth_pair(dev1_name, dev2_name, mtu=None):
     """Create a pair of veth devices with the specified names,
     deleting any previous devices with those names.
     """
@@ -1352,7 +1352,7 @@ def _create_veth_pair(dev1_name, dev2_name):
         utils.execute('ip', 'link', 'set', dev, 'up', run_as_root=True)
         utils.execute('ip', 'link', 'set', dev, 'promisc', 'on',
                       run_as_root=True)
-        _set_device_mtu(dev)
+        _set_device_mtu(dev, mtu)
 
 
 def _ovs_vsctl(args):
@@ -1365,7 +1365,7 @@ def _ovs_vsctl(args):
         raise exception.AgentError(method=full_args)
 
 
-def create_ovs_vif_port(bridge, dev, iface_id, mac, instance_id):
+def create_ovs_vif_port(bridge, dev, iface_id, mac, instance_id, mtu=None):
     _ovs_vsctl(['--', '--if-exists', 'del-port', dev, '--',
                 'add-port', bridge, dev,
                 '--', 'set', 'Interface', dev,
@@ -1373,12 +1373,13 @@ def create_ovs_vif_port(bridge, dev, iface_id, mac, instance_id):
                 'external-ids:iface-status=active',
                 'external-ids:attached-mac=%s' % mac,
                 'external-ids:vm-uuid=%s' % instance_id])
-    _set_device_mtu(dev)
+    _set_device_mtu(dev, mtu)
 
 
-def delete_ovs_vif_port(bridge, dev):
+def delete_ovs_vif_port(bridge, dev, delete_dev=True):
     _ovs_vsctl(['--', '--if-exists', 'del-port', bridge, dev])
-    delete_net_dev(dev)
+    if delete_dev:
+        delete_net_dev(dev)
 
 
 def ovs_set_vhostuser_port_type(dev):
@@ -1411,6 +1412,20 @@ def create_tap_dev(dev, mac_address=None):
                           run_as_root=True, check_exit_code=[0, 2, 254])
         utils.execute('ip', 'link', 'set', dev, 'up', run_as_root=True,
                       check_exit_code=[0, 2, 254])
+
+
+def create_fp_dev(dev, sockpath, sockmode):
+    if not device_exists(dev):
+        utils.execute('fp-vdev', 'add', dev, '--sockpath', sockpath,
+                      '--sockmode', sockmode, run_as_root=True)
+        _set_device_mtu(dev)
+        utils.execute('ip', 'link', 'set', dev, 'up', run_as_root=True,
+                    check_exit_code=[0, 2, 254])
+
+
+def delete_fp_dev(dev):
+    if device_exists(dev):
+        utils.execute('fp-vdev', 'del', dev, run_as_root=True)
 
 
 def delete_net_dev(dev):
@@ -1948,8 +1963,6 @@ class NeutronLinuxBridgeInterfaceDriver(LinuxNetInterfaceDriver):
         bridge = self.BRIDGE_NAME_PREFIX + str(network['uuid'][0:11])
         return bridge
 
-# provide compatibility with existing configs
-QuantumLinuxBridgeInterfaceDriver = NeutronLinuxBridgeInterfaceDriver
 
 iptables_manager = IptablesManager()
